+++ TrainTrain/SeatJsonPoco.cs
@@ -1,0 +1,9 @@
+﻿namespace TrainTrain
+{
+    public class SeatJsonPoco
+    {
+        public string booking_reference { get; set; }
+        public string seat_number { get; set; }
+        public string coach { get; set; }
+    }
+}
