+++ TrainTrain/TrainTrain.csproj
@@ -49,6 +49,7 @@
     <Compile Include="BookingReferenceService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Seat.cs" />
+    <Compile Include="SeatJsonPoco.cs" />
     <Compile Include="ThreasholdManager.cs" />
     <Compile Include="Train.cs" />
     <Compile Include="TrainCaching.cs" />
